requirements.txt
# ===================================================================
# artellapipe-launcher-plugins-dccselector requirements file
# ===================================================================
Qt.py
pytest
pytest-cov
coveralls
sphinx
sphinx_rtd_theme
python-json-logger
sentry-sdk
tpDcc-libs-python
tpDcc-core
tpDcc-libs-qt
artellapipe
artellapipe-launcher

